refactor(learner): migrate `useIsWithinViewport` to class-based reactive state (`IsWithinViewport`)
