+++ apps/learner/src/lib/helpers/index.ts
@@ -1,1 +1,1 @@
-export * from './use-is-within-viewport.svelte.js';
+export * from './is-within-viewport.svelte.js';
